requirements.txt
firebase-admin==2.17.0
google-cloud==0.34.0
requests==2.22.0
